pyproject.toml
[project]
name = "smart-gmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "celery>=5.5.3",
    "fastapi>=0.116.1",
    "fastmcp>=2.11.3",
    "golf-mcp>=0.2.0",
    "jinja2>=3.1.6",
    "mcp>=1.13.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "schedule>=1.2.2",
    "streamlit>=1.48.1",
    "uvicorn[standard]>=0.35.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.9",
]
